+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 * `convert/from_seurat_to_h5mu`: Converts a Seurat object to a MuData object (PR #1078, #1079, #1082).
 
+* `annotate/celltypist`: Enable CUDA acceleration for CellTypist annotation (PR #1091).
+
 ## EXPERIMENTAL
 
 * `differential_expression/deseq2`: Performs differential expression analysis using DESeq2 on bulk or pseudobulk datasets (PR #1044).